+++ pioreactor/background_jobs/growth_rate_calculating.py
@@ -396,6 +396,7 @@ class GrowthRateCalculator(BackgroundJob):
         od_readings = decode(message.payload, type=structs.ODReadings)
 
         self.update_state_from_observation(od_readings)
+
         return
 
     def update_state_from_observation(
@@ -485,7 +486,8 @@ class GrowthRateCalculator(BackgroundJob):
     def respond_to_dosing_event(self, dosing_event: structs.DosingEvent) -> None:
         # here we can add custom logic to handle dosing events.
         # an improvement to this: the variance factor is proportional to the amount exchanged.
-        self.update_ekf_variance_after_event(minutes=0.40, factor=2500)
+        if dosing_event.event != "remove_waste":
+            self.update_ekf_variance_after_event(minutes=0.40, factor=2500)
 
     def start_passive_listeners(self) -> None:
         # process incoming data