+++ pioreactor/cli/pio.py
@@ -290,7 +290,7 @@ def version(verbose: bool) -> None:
         click.echo(f"Image version:          {whoami.get_image_git_hash()}")
         if whoami.am_I_leader():
             try:
-                result = get("http://127.0.0.1/api/ui_version")
+                result = get("http://127.0.0.1/api/versions/ui")
                 result.raise_for_status()
                 ui_version = result.body.decode()
             except Exception:
@@ -464,9 +464,12 @@ def update_app(branch: Optional[str], source: Optional[str], version: Optional[s
 
     else:
         tag = get_tag_to_install(version)
-        release_metadata = loads(
-            get(f"https://api.github.com/repos/pioreactor/pioreactor/releases/{tag}").body
-        )
+        response = get(f"https://api.github.com/repos/pioreactor/pioreactor/releases/{tag}")
+        if response.raise_for_status():
+            logger.error(f"Version {version} not found")
+            raise click.Abort()
+
+        release_metadata = loads(response.body)
         version_installed = release_metadata["tag_name"]
         for asset in release_metadata["assets"]:
             # add the following files to the release. They should ideally be idempotent!
@@ -624,7 +627,7 @@ if whoami.am_I_leader():
     run.add_command(jobs.watchdog.click_watchdog)
     run.add_command(actions.export_experiment_data.click_export_experiment_data)
     run.add_command(actions.backup_database.click_backup_database)
-    run.add_command(actions.execute_experiment_profile.click_execute_experiment_profile)
+    run.add_command(actions.experiment_profile.click_experiment_profile)
 
     @pio.command(short_help="access the db CLI")
     def db() -> None: