+++ pioreactor/tests/test_experiment_profile_structs.py
@@ -15,17 +15,25 @@ experiment_profile_name: minimal
 
 def test_simple1():
     file = """
-experiment_profile_name: test_simple1
+experiment_profile_name: demo_stirring_example
 
 metadata:
-  author: Jane Doe
-  description:
+  author: Cam Davidson-Pilon
+  description: A simple profile to start stirring in your Pioreactor(s), update RPM at 90 seconds, and turn off after 180 seconds.
 
 common:
-  od_reading:
+  stirring:
     actions:
       - type: start
-        hours_elapsed: 1.0
+        hours_elapsed: 0.0
+        options:
+          target_rpm: 400.0
+      - type: update
+        hours_elapsed: 0.025
+        options:
+          target_rpm: 800.0
+      - type: stop
+        hours_elapsed: 0.05
 """
     assert decode(file, type=structs.Profile) is not None
 
@@ -97,7 +105,7 @@ common:
     actions:
       - type: start
         hours_elapsed: 0.0
-        parameters:
+        options:
           target_rpm: 200.0
       - type: stop
         hours_elapsed: 2.0
@@ -111,80 +119,34 @@ pioreactors:
 
 def test_complex():
     file = """
-experiment_profile_name: multi_jobs_dependencies
+experiment_profile_name: complex_example
 
 metadata:
-  author: John Doe
-  description: Multiple jobs with dependencies, aliases, and plugin requirements in bioreactors
-  organism_used: yeast
-
-plugins:
-  - name: dosing_plugin
-    version: ">=1.2.0"
-
-labels:
-  bioreactor_A: BR-001
-  bioreactor_B: BR-002
+  author: Cam Davidson-Pilon
+  description: A more complex profile to start stirring, heating, and (later) od_reading and growth_rate_calculating.
 
 common:
+  stirring:
+    actions:
+      - type: start
+        hours_elapsed: 0.0
+        options:
+          target_rpm: 400.0
+  temperature_control:
+    actions:
+      - type: start
+        hours_elapsed: 0.0
+        options:
+          automation_name: thermostat
+          target_temperature: 30
   od_reading:
     actions:
       - type: start
-        hours_elapsed: 0.5
-      - type: pause
-        hours_elapsed: 2.0
-      - type: resume
-        hours_elapsed: 2.5
-      - type: stop
-        hours_elapsed: 5.0
-
-pioreactors:
-  bioreactor_A:
-    jobs:
-      stirring:
-        actions:
-          - type: start
-            hours_elapsed: 0.0
-            options:
-              target_rpm: 200.0
-          - type: update
-            hours_elapsed: 1.0
-            options:
-               target_rpm: 250.0
-          - type: stop
-            hours_elapsed: 5.0
-      dosing_control:
-        actions:
-          - type: start
-            hours_elapsed: 1.0
-            options:
-              automation_name: glucose_dosing
-              glucose_rate: 5
-          - type: stop
-            hours_elapsed: 5.0
-  bioreactor_B:
-    jobs:
-      stirring:
-        actions:
-          - type: start
-            hours_elapsed: 0.0
-            options:
-              target_rpm: 250.0
-          - type: update
-            hours_elapsed: 2.0
-            options:
-              target_rpm: 300.0
-          - type: stop
-            hours_elapsed: 5.0
-      dosing_control:
-        actions:
-          - type: start
-            hours_elapsed: 1.0
-            options:
-              automation_name: nitrogen_dosing
-              nitrogen_rate: 10
-          - type: stop
-            hours_elapsed: 5.0
+        hours_elapsed: 0.25
+  growth_rate_calculating:
+    actions:
+      - type: start
+        hours_elapsed: 0.33
 """
     assert decode(file, type=structs.Profile) is not None