+++ pioreactor/tests/test_temperature_control.py
@@ -29,14 +29,14 @@ def test_thermostat_automation() -> None:
     ) as algo:
         pause(2)
 
-        # 55 is too high - clamps to 50
+        # 55 is too high - clamps to MAX_TARGET_TEMP
         pubsub.publish(
             f"pioreactor/{unit}/{experiment}/temperature_automation/target_temperature/set",
             55,
         )
         pause(2)
 
-        assert algo.automation_job.target_temperature == 50
+        assert algo.automation_job.target_temperature == algo.automation_job.MAX_TARGET_TEMP
 
         pubsub.publish(
             f"pioreactor/{unit}/{experiment}/temperature_automation/target_temperature/set",
@@ -121,7 +121,7 @@ def test_heating_is_reduced_when_set_temp_is_exceeded() -> None:
         pause()
         assert t.heater_duty_cycle == 50
         pause()
-        t.read_external_temperature_and_check_temp()
+        t.read_external_temperature()
         pause()
 
         assert 0 < t.heater_duty_cycle < 50
@@ -153,7 +153,7 @@ def test_heating_stops_when_max_temp_is_exceeded() -> None:
         assert t.heater_duty_cycle == 50
         pause()
         pause()
-        t.read_external_temperature_and_check_temp()
+        t.read_external_temperature()
         pause()
         pause()
 
@@ -682,7 +682,7 @@ def test_temperature_control_and_thermostats_relationship() -> None:
         pause()
 
         # run evaluate_and_publish_temperature, this locks the PWM from anyone updating it directly.
-        thread = threading.Thread(target=tc.evaluate_temperature, daemon=True)
+        thread = threading.Thread(target=tc.infer_temperature, daemon=True)
         thread.start()
         pause()