+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### 23.6.7
+
+#### Highlights
+ - Support for viewing, starting and stopping _experiment profiles_ in the UI!
+ - Adding manual dosing adjustment form under Dosing tab in the UI!
+
+#### Everything else
+ - New API for experiment profiles: `pio run experiment_profile`, with subcommands `execute` and `verify`. So what use to be `pio run execute_experiment_profile <filename>` is now: `pio run experiment_profile execute <filename>`. The `verify` subcommand is for checking the yaml file for errors.
+ - new leader CLI command: `pios cp <filepath>` will move a file on your leader to the entire cluster. This is useful for distributing plugins and Python wheels across your workers.
+ - plugins can now add `post_install.sh` and `pre_uninstall.sh` bash scripts.
+ - added `[stirring]` option `duration_between_updates_seconds` to config, default is 23.0.
+ - PIDMorbidostat has a configuration parameter `[dosing_automation.pid_morbidostat].minimum_dosing_volume_ml` (default 0.1). If a calculated volume to be dosed is less than this parameter, then it's set to 0.0 instead.
+ - adding `--manually` flag to pump actions, ex: `pio run add_media --ml 1 --manually`. This _doesn't_ run the pump, but still fires a dosing event, which downstream jobs listen to (ex: saves to database, will update metrics). See next change:
+
 ### 23.5.16
  - UX improvements to `pio run pump_calibration`
  - `monitor` is more robust, so as to give users better access to information instead of hard-failing.
@@ -15,7 +29,6 @@ Also shipping this version is early support for experiment profiles. What are th
 For examples of yaml files, see the repo: https://github.com/Pioreactor/experiment_profile_examples
 
 To use a profile, save the yaml file to your leader Pioreactor. Run it with `pio run execute_experiment_profile <path_to_yaml>`. Note that killing the `execute_experiment_profile` will only stop execution of upcoming actions, and won't stop any jobs that have already started from the profile.
-
 
 
 ### 23.4.28